Scripts fixed
This fix corrects the import bug
